requirements.txt
Django==5.2
django-watson==1.6.2
django-phonenumber-field==8.0.0
phonenumbers==8.13.30
PyMySQL==1.1.0
requests==2.31.0
pytz==2023.3
